Python/VirtualPT_Server/requirements.txt
Flask==3.0.3
boto3==1.34.136
flask-cors==4.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1